# [inductor] Lower torch.exp2 and use it for torch.pow(2, x)