--- a/torch/_refs/__init__.py
+++ b/torch/_refs/__init__.py
@@ -1097,6 +1097,12 @@ def pow(
             return a * a  # type: ignore[return-value]
         elif b == 0.5:
             return torch.sqrt(a)  # type: ignore[arg-type]
+    elif isinstance(a, Number):
+        if a == 1.0:
+            return torch.fill(b, True)
+        if a == 2.0 and (utils.is_float_dtype(b.dtype) or utils.is_complex_dtype(b.dtype)):
+            return torch.exp2(b)
+
     return prims.pow(a, b)