@ngimel 2023-01-20
exp2 is a prim (or should be) so using primTorch for it is kinda irrelevant.