@lezcano 2023-01-20
My point is that we have pretty much all this logic repeated here and in the ref for `pow` because of that broadcasting issue :/