@ngimel 2023-01-20
> all this logic repeated here and in the ref for pow because of that broadcasting issue :/
Not quite, having different pow lowerings in primtorch and inductor is reasonable IMO. Backends that can just write `::pow(a,b)` (most of them) are in a different position than inductor that has to do a lot of gymnastics for integer `pow`.