@peterbell10 2023-01-20
> Backends that can just write `::pow(a,b)` (most of them) are in a different position than inductor that has to do a lot of gymnastics for integer pow.

In general it would be nice if  backend-agnostic optimizations happened at the decomposition level and inductor-specific gymnastics can happen inside the `prim`. So all of the `pow_recursive` stuff would stay in the lowering for `prims.pow` while dispatching to a more specific prim like `prims.sqrt` for `pow(x, 0.5)` can be done by the decomposition.

This would also mitigate the issue of the reference and lowering falling out of sync. e.g. For `ceil` the lowering is short-circuited for integers:
```python
def ceil(x):
    if is_integer_type(x):
        return x
    fn = ops_wrapper("ceil")
    return make_pointwise(fn)(x)
```

but the reference isn't
```python
def ceil(a):
    return prims.ceil(a)
```